Internal tidying: move more implementation details out of mod.rs files.
To make mod.rs files tidier, move some implementation details out of
mod.rs files and into the underlying source files.

This is also a first step towards #250.
