+++ src/fs/statx.rs
@@ -2,7 +2,9 @@
 
 use crate::{imp, io, path};
 use imp::fd::AsFd;
-use imp::fs::{AtFlags, Statx};
+use imp::fs::AtFlags;
+
+pub use imp::fs::Statx;
 
 /// `STATX_*` constants.
 pub use imp::fs::StatxFlags;
