+++ src/time/mod.rs
@@ -1,7 +1,5 @@
 //! Time-related operations.
 
-use crate::imp;
-
 mod clock;
 #[cfg(any(target_os = "android", target_os = "fuchsia", target_os = "linux"))]
 mod timerfd;
@@ -12,7 +10,7 @@ mod timerfd;
 pub use clock::clock_getres;
 #[cfg(not(target_os = "wasi"))]
 pub use clock::{clock_gettime, clock_gettime_dynamic, ClockId, DynamicClockId};
-pub use imp::time::{Nsecs, Secs, Timespec};
+pub use clock::{Nsecs, Secs, Timespec};
 #[cfg(any(target_os = "android", target_os = "fuchsia", target_os = "linux"))]
 pub use timerfd::{
     timerfd_create, timerfd_gettime, timerfd_settime, Itimerspec, TimerfdClockId, TimerfdFlags,
